refactor(frontend): extract MatchCard from identity page (505 -> 179 lines)
